I'm clueless
really I am, I really thought that what was done in help to me, that
worked, that I could take it and run with it, and yet because I'm trying
something new, again, apparently, it's like I've stepped into yet
another circle that is of itself twice wise inside the other two that
I've had help going through. How much farther to the center of this
ancient power circle must I go before the knowledge which I seek and
desire is put upon me such that I may know it?